Rename sqrt function to squareRoot and update return type to double

+++ library-practice/library.h
@@ -8,6 +8,6 @@ int subtract(int x, int y);
 int multiply(int x, int y);
 float divide(int x, int y);
 int square(int x);
-int squareRoot(int x);
+double squareRoot(int x);
 
 #endif // basic_calcu_lib_h
